remove code repeating
