+++ src/components/Messages/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import styled from 'styled-components';
 import ScrollBar from 'react-scrollbars-custom';
-import { uniqueId } from 'lodash';
 import moreOptionSrc from '../../img/icons/chat-more-options.svg';
 import Messages from '../../common/chat/messages';
 import SubmitMessage from '../../common/chat/submit-message';
 import PageSearchInput from '../../common/Inputs/PageSearch';
 import PageWrapper from '../../common/pageWrapper';
+import { groupMessagesByUser } from '../../common/chat/helper';
 
 const Wrapper = styled.div`
   max-width: 1456px;
@@ -203,271 +203,6 @@ const ContentUserImg = styled.img`
 `;
 
 const scrollBarStyles = { width: '100%', height: '100%', paddingRight: 10 };
-
-// добавить сообщениям id либо id придёт от сервера
-// отсортировать по дате либо придут отсортированные
-const testData = [
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '12:55',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '13:40',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '13:40',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '12:55',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '13:40',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '13:40',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '12:55',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Nulla porttitor ligula quam, quis auctor felis consectetur non',
-    dateSend: '05/06/20',
-    timeSend: '12:59',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '13:40',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'bogdan13',
-    name: 'Богдан',
-    messages: 'Pellentesque blandit nibh at leo venenatis, in semper ipsum dictum',
-    dateSend: '05/06/20',
-    timeSend: '12:58',
-    image: 'https://igate.com.ua/upload/photo/0001/0001/3383/6955/55.jpg',
-  },
-  {
-    id: uniqueId(),
-    username: 'kirill22',
-    name: 'Кирилл',
-    messages: 'Lorem! Aliquam erat volutpat?',
-    dateSend: '05/06/20',
-    timeSend: '13:40',
-    image: 'https://st.kp.yandex.net/images/actor_iphone/iphone360_1746394.jpg',
-  },
-];
-
-const groupMessagesByUser = testData.reduce((acc, el) => {
-  const lastAccElement = acc[acc.length - 1];
-  const lastAccIndex = acc.length - 1;
-
-  const initialMessagesArray = { ...el, messages: [{ title: el.messages, id: uniqueId() }] };
-  if (lastAccElement === undefined) return [initialMessagesArray];
-
-  if (lastAccElement.username === el.username) {
-    const newObject = {
-      ...acc[lastAccIndex],
-      messages: [...acc[lastAccIndex].messages, { title: el.messages, id: uniqueId() }],
-    };
-    return [...acc.splice(0, lastAccIndex), newObject];
-  }
-  return [...acc, initialMessagesArray];
-}, []);
 
 const MessagesPage = () => {
   const renderMessages = () => groupMessagesByUser.map((el) => {
